Make the program file
oh man this sucks. I had file with a working rgb function but then windows went kamikaze and I lost it, so this is my 5 minute recreation that obviosly doesn't work. All the includes defines and what not were taken from the ili9341 graphicstest or something similar to that name, then I stripped it of anything I didn't need, and started building my program off of that.