again a simple but very good upgrade
little history on the rendering: started out rendering the whole board each frame, this took about 1 second to render. then i changed to rendering the whole snake, and the position behind it to remove the rendered tail this would slow down overtime, and if the snake got long enough to fill the whole board it would be taking 1 second to render again. Now it only renders the head, the position behind the head, and the white space behind the tail. this won't slow down overtime since it's only rendering the same amount of stuff each time. Also on the movement, it used to use a board with the directions that the snake went in to see where each piece would go. I'm sure this could work if i didn't need the ghost to unrender the tail, but it caused issues with that, so now each piece just moves to the one before it in the array which works flawlessly, and also those 2 board that held the snake, and the snake directions took up about 20% of dynamic memory each which is a lot. now it's only using 55% of dynamic memory for the entire system which is very good.